Move image-orientation property outside of mako
Fixes https://github.com/servo/servo/issues/19765
